Rolling # Animation - Rolling 애니메이션에 무적 이벤트
# Animator
- 구르는 동안 플레이어를 무적으로 만들어주는 bool 매개변수

# AnimatorHandler
- Rolling 애니메이션에 이벤트로 추가할 함수

# PlayerManager
- 플레이어가 현재 무적인지 아닌지를 나타내는 bool 변수
- Update 내에서 Animator의 무적 bool 매개변수로 해당 bool 변수 관리
- 스테미너 회복

# PlayerStats + Others
- 공격력, HP, Stamina 관련 변수 전부 float형으로 교체
- 스테미너 회복함수, 행동중에는 회복 안되도록

+++ Assets/Scripts/AI/EnemyStats.cs
@@ -14,12 +14,12 @@ namespace sg {
             currentHealth = maxHealth;
         }
 
-        private int SetMaxHealthFromHealthLevel() {
+        private float SetMaxHealthFromHealthLevel() {
             maxHealth = healthLevel * 10;
             return maxHealth;
         }
 
-        public void TakeDamage(int damage) {
+        public void TakeDamage(float damage) {
             if (isDead) return;
             currentHealth -= damage;
             animator.Play("Damage");
